requirements.txt
opencv-python==4.7.0.72
pdf2image==1.16.3
Pillow==9.5.0
pytesseract==0.3.10
reportlab==3.6.13
